refactor: [move-only] Merge core_io module
This can be reviewed with the git option
--color-moved=dimmed-zebra
